fix: separate template processor from base theme
